requirements.txt

fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
Jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
fastapi-mail==1.4.1
reportlab==4.2.5
